requirements.txt
flask
google-generativeai
python-dotenv
beautifulsoup4
requests